requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
joblib==1.5.1
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-decouple==3.8
python-dotenv==1.1.1
python-owasp-zap-v2.4==0.1.0
reportlab==4.4.4
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
xmltodict==1.0.2
zaproxy==0.4.0
